exalab/producer/requirements.txt
kafka_python==2.0.2
requests==2.22.0
